+++ internal/kv/tikv/txn_tikv.go
@@ -772,8 +772,7 @@ func convertEmptyStringToByte(value string) ([]byte, error) {
 	if len(value) == 0 {
 		return EmptyValueByte, nil
 	} else if value == EmptyValueString {
-		return nil, fmt.Errorf("Value for key is reserved by EmptyValue: %s", EmptyValueString)
-	} else {
-		return []byte(value), nil
+		return nil, fmt.Errorf("value for key is reserved by EmptyValue: %s", EmptyValueString)
 	}
+	return []byte(value), nil
 }
